push fixes to rate.

+++ src/com/skilldistillery/foodtruck/app/FoodTruck.java
@@ -42,7 +42,7 @@ public class FoodTruck {
 		this.foodType = foodType;
 	}
 
-	public double getRating() {
+	public int getRating() {
 		return rating;
 	}
